Move gtc_backend.base to gtc_backend.common
